Problem: rigidness of bgworker-based extensions
There are some extensions that require a worker to function.

Realistically speaking, we need to make these extensions
shared_preloaded_library in postgres config and we need to restart
PostgreSQL every time we want to change it.

Solution: make one "manager" extension to be a preloaded library

This extension (pgextkit) would be responsible for providing shared
memory resources and run background workers for extensions, whether at a
startup time or during runtime.

This commit is an early draft that seems to work.